requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==21.2.0
awscli==1.22.34
bcrypt==3.2.0
blinker==1.4
botocore==1.23.34
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.17.1
duplicity==0.8.21
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.2
fasteners==0.14.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
future==0.18.2
greenlet==3.0.3
h11==0.14.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.4
jmespath==0.10.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libevdev==0.5
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.9.2
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
mpmath==1.3.0
netifaces==0.11.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
pillow==10.3.0
protobuf==3.12.4
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
PyAutoGUI==0.9.54
pycairo==1.20.1
pycups==2.0.1
pydantic==2.8.2
pydantic_core==2.20.1
PyGetWindow==0.0.9
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.18.1
PyScreeze==0.1.30
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.9
python3-xlib==0.15
pytweening==1.2.0
pytz==2022.1
pyudev==0.22.0
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
roman==3.3
rsa==4.8
s3transfer==0.5.0
scipy==1.14.1
seaborn==0.13.2
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.38.2
sympy==1.13.2
systemd-python==234
texttable==1.6.4
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ultralytics==8.2.84
ultralytics-thop==2.0.5
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.30.3
vboxapi==1.0
wadllib==1.3.6
websocket-client==1.2.3
xdg==5
xkit==0.0.0
zipp==1.0.0
